+++ tests/ResultTest.php
@@ -17,7 +17,7 @@ final class ResultTest extends \PHPUnit_Framework_TestCase
             ->method('getSortCode')
             ->will($this->returnValue($sortCode));
 
-        $result = new Result($account, true, true);
+        $result = new Result($account, true, true, self::now());
 
         $this->assertSame($sortCode, $result->getSortCode());
     }
@@ -32,7 +32,7 @@ final class ResultTest extends \PHPUnit_Framework_TestCase
             ->method('getAccountNumber')
             ->will($this->returnValue($accountNumber));
 
-        $result = new Result($account, true, true);
+        $result = new Result($account, true, true, self::now());
 
         $this->assertSame($accountNumber, $result->getAccountNumber());
     }
@@ -41,10 +41,10 @@ final class ResultTest extends \PHPUnit_Framework_TestCase
     {
         $account = $this->getMockForAbstractClass('Cs278\BankModulus\BankAccountInterface');
 
-        $result = new Result($account, true, true);
+        $result = new Result($account, true, true, self::now());
         $this->assertTrue($result->isValidated());
 
-        $result = new Result($account, false, null);
+        $result = new Result($account, false, null, self::now());
         $this->assertFalse($result->isValidated());
     }
 
@@ -52,14 +52,14 @@ final class ResultTest extends \PHPUnit_Framework_TestCase
     {
         $account = $this->getMockForAbstractClass('Cs278\BankModulus\BankAccountInterface');
 
-        $result = new Result($account, true, true);
+        $result = new Result($account, true, true, self::now());
         $this->assertTrue($result->isValid());
 
-        $result = new Result($account, true, false);
+        $result = new Result($account, true, false, self::now());
         $this->assertFalse($result->isValid());
 
         // Test unvalidated.
-        $result = new Result($account, false, null);
+        $result = new Result($account, false, null, self::now());
         $this->assertTrue($result->isValid());
         $this->assertTrue($result->isValid(true));
         $this->assertFalse($result->isValid(false));
@@ -70,7 +70,7 @@ final class ResultTest extends \PHPUnit_Framework_TestCase
     {
         $account = $this->getMockForAbstractClass('Cs278\BankModulus\BankAccountInterface');
 
-        $result = new Result($account, false, null);
+        $result = new Result($account, false, null, self::now());
 
         try {
             $result->isValid($value);
@@ -87,13 +87,82 @@ final class ResultTest extends \PHPUnit_Framework_TestCase
         $this->fail('Failed to catch exception');
     }
 
+    public function testGetValidated()
+    {
+        $expected = self::now();
+        $account = $this->getMockForAbstractClass('Cs278\BankModulus\BankAccountInterface');
+        $result = new Result($account, true, false, $expected);
+
+        $this->assertSame($expected, $result->getValidatedAt());
+    }
+
+    /**
+     * @requires function error_clear_last
+     * @group legacy
+     */
+    public function testConstructorWithoutValidatedAt()
+    {
+        error_clear_last();
+
+        $account = $this->getMockForAbstractClass('Cs278\BankModulus\BankAccountInterface');
+        $result = new Result($account, true, false, null);
+        $error = error_get_last();
+
+        $this->assertInstanceOf('Cs278\\BankModulus\\Result', $result);
+        $this->assertArraySubset([
+            'message' => '$validatedAt will become a required argument of Cs278\\BankModulus\\Result::__construct() in version 2.0.0.',
+            'type' => E_USER_DEPRECATED,
+        ], $error);
+
+        $this->assertInstanceOf(get_class(self::now()), $result->getValidatedAt());
+    }
+
+    public function testConstructorWithInvalidValidatedAt()
+    {
+        $account = $this->getMockForAbstractClass('Cs278\BankModulus\BankAccountInterface');
+
+        try {
+            new Result($account, false, null, new \stdClass());
+        } catch (\Exception $e) {
+            if ($e instanceof \PHPUnit_Exception) {
+                throw $e;
+            }
+
+            $this->assertInvalidArgumentException('validatedAt should be an instance of DateTimeImmutable, DateTime or null, got: `%s`', $e);
+
+            return;
+        }
+
+        $this->fail('Failed to catch exception');
+    }
+
+    /**
+     * @requires function error_clear_last
+     * @requires PHP 7.1.0
+     */
+    public function testConstructorWithDeprecatedValidatedAt()
+    {
+        error_clear_last();
+
+        $account = $this->getMockForAbstractClass('Cs278\BankModulus\BankAccountInterface');
+        $result = new Result($account, false, null, new \DateTime());
+
+        $error = error_get_last();
+
+        $this->assertInstanceOf('Cs278\\BankModulus\\Result', $result);
+        $this->assertArraySubset([
+            'message' => '$validatedAt argument of Cs278\BankModulus\Result::__construct() will require a DateTimeImmutable instance in version 2.0.0.',
+            'type' => E_USER_DEPRECATED,
+        ], $error);
+    }
+
     /** @dataProvider dataNonBooleans */
     public function testConstructorSpecKnownValidation($value)
     {
         $account = $this->getMockForAbstractClass('Cs278\BankModulus\BankAccountInterface');
 
         try {
-            new Result($account, $value, true);
+            new Result($account, $value, true, self::now());
         } catch (\Exception $e) {
             if ($e instanceof \PHPUnit_Exception) {
                 throw $e;
@@ -113,7 +182,7 @@ final class ResultTest extends \PHPUnit_Framework_TestCase
         $account = $this->getMockForAbstractClass('Cs278\BankModulus\BankAccountInterface');
 
         try {
-            new Result($account, true, $value);
+            new Result($account, true, $value, self::now());
         } catch (\Exception $e) {
             if ($e instanceof \PHPUnit_Exception) {
                 throw $e;
@@ -133,7 +202,7 @@ final class ResultTest extends \PHPUnit_Framework_TestCase
         $account = $this->getMockForAbstractClass('Cs278\BankModulus\BankAccountInterface');
 
         try {
-            new Result($account, false, $value);
+            new Result($account, false, $value, self::now());
         } catch (\Exception $e) {
             if ($e instanceof \PHPUnit_Exception) {
                 throw $e;
@@ -180,5 +249,13 @@ final class ResultTest extends \PHPUnit_Framework_TestCase
         $this->assertInstanceOf('Cs278\BankModulus\Exception\InvalidArgumentException', $e);
         $this->assertInstanceOf('InvalidArgumentException', $e);
         $this->assertStringMatchesFormat($expectedMessage, $e->getMessage());
+    }
+
+    /** @return \DateTimeImmutable|\DateTime */
+    private static function now()
+    {
+        return class_exists('DateTimeImmutable')
+            ? new \DateTimeImmutable()
+            : new \DateTime();
     }
 }